requirements.txt
PyQt6>=6.5
pyserial>=3.5
pyqtgraph>=0.13
numpy>=1.24
scipy>=1.11
openai>=1.0
python-dotenv>=1.0
PyOpenGL>=3.1
